+++ tests/graph_test.py
@@ -1,7 +1,7 @@
 from mythril.analysis.callgraph import generate_graph
 from mythril.analysis.symbolic import SymExecWrapper
 from mythril.ethereum import util
-from mythril.solidity.soliditycontract import ETHContract
+from mythril.solidity.soliditycontract import EVMContract
 from tests import *
 import re
 
@@ -17,7 +17,7 @@ class GraphTest(BaseTestCase):
                 input_file.name + ".graph.html"
             )
 
-            contract = ETHContract(input_file.read_text())
+            contract = EVMContract(input_file.read_text())
 
             sym = SymExecWrapper(
                 contract,
