+++ tests/report_test.py
@@ -2,7 +2,7 @@ from mythril.analysis.report import Report
 from mythril.analysis.security import fire_lasers
 from mythril.analysis.symbolic import SymExecWrapper
 from mythril.ethereum import util
-from mythril.solidity.soliditycontract import ETHContract
+from mythril.solidity.soliditycontract import EVMContract
 from multiprocessing import Pool, cpu_count
 import pytest
 import json
@@ -22,7 +22,7 @@ def _fix_debug_data(json_str):
 
 
 def _generate_report(input_file):
-    contract = ETHContract(input_file.read_text(), enable_online_lookup=False)
+    contract = EVMContract(input_file.read_text(), enable_online_lookup=False)
     sym = SymExecWrapper(
         contract,
         address=(util.get_indexed_address(0)),
