+++ tests/svm_test.py
@@ -2,7 +2,7 @@ import json
 from mythril.analysis.security import get_detection_module_hooks
 from mythril.analysis.symbolic import SymExecWrapper
 from mythril.analysis.callgraph import generate_graph
-from mythril.ethereum.ethcontract import ETHContract
+from mythril.ethereum.evmcontract import EVMContract
 from mythril.solidity.soliditycontract import SolidityContract
 
 from mythril.laser.ethereum.state.account import Account
@@ -105,7 +105,7 @@ class SVMTestCase(BaseTestCase):
 
         code = "0x60606040525b603c5b60006010603e565b9050593681016040523660008237602060003683856040603f5a0204f41560545760206000f35bfe5b50565b005b73c3b2ae46792547a96b9f84405e36d0e07edcd05c5b905600a165627a7a7230582062a884f947232ada573f95940cce9c8bfb7e4e14e21df5af4e884941afb55e590029"
 
-        contract = ETHContract(code)
+        contract = EVMContract(code)
         sym = SymExecWrapper(contract, "0xd0a6E6C543bC68Db5db3A191B171A77407Ff7ccf")
 
         html = generate_graph(sym)
